persist data

+++ PhucLocBao/src/main/java/vn/com/phuclocbao/dao/CompanyDao.java
@@ -3,5 +3,5 @@ package vn.com.phuclocbao.dao;
 import vn.com.phuclocbao.entity.CompanyEntity;
 
 public interface CompanyDao{
-	public CompanyEntity findById(Long id);
+	public CompanyEntity findById(Integer id);
 }
